Move all journey tests to `journeys` package
