+++ .github/workflows/run-percy.yml
@@ -41,7 +41,7 @@ jobs:
     - name: Grant execute permission for gradlew
       run: chmod +x gradlew
     - name: Run journey test and send to Percy
-      run: percy exec -- ./gradlew clean test --tests org.codeforamerica.shiba.newjourneys.FullFlowJourneyTest.fullApplicationWithDocumentUploads
+      run: percy exec -- ./gradlew clean test --tests org.codeforamerica.shiba.journeys.FullFlowJourneyTest.fullApplicationWithDocumentUploads
       env:
         PERCY_TOKEN: ${{ secrets.PERCY_TOKEN }}
         S3_BUCKET: ${{ secrets.S3_TEST_BUCKET }}
