+++ src/test/java/org/codeforamerica/shiba/journeys/DocumentUploadJourneyTest.java
@@ -1,6 +1,5 @@
-package org.codeforamerica.shiba.newjourneys;
+package org.codeforamerica.shiba.journeys;
 
-import org.codeforamerica.shiba.pages.journeys.JourneyTest;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.By;
